requirements.txt
streamlit==1.28.0
tensorflow==2.14.0
opencv-python==4.8.1.75
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0
